requirements.txt
Pillow==8.3.1
pytesseract==0.3.8
tesseract==0.1.3
